requirements.txt
-i https://pypi.org/simple
certifi==2019.11.28
chardet==3.0.4
idna==2.9
psycopg2-binary==2.8.4
pytelegrambotapi==3.6.7
requests==2.23.0
six==1.14.0
urllib3==1.25.8
